+++ src/guards/string.guard.ts
@@ -122,6 +122,7 @@ export class StringGuard extends Guard<StringRule> {
      * @remarks Chainable method.
      * @param alphabet - {@link https://en.wikipedia.org/wiki/ISO_basic_Latin_alphabet | 'basic-latin'} |
      * {@link https://en.wikipedia.org/wiki/List_of_precomposed_Latin_characters_in_Unicode | 'precomposed-latin'} |
+     * {@link https://en.wikipedia.org/wiki/Danish_orthography | 'dan'} |
      * {@link https://en.wikipedia.org/wiki/German_orthography | 'deu'} |
      * {@link https://en.wikipedia.org/wiki/Finnish_orthography | 'fin'} |
      * {@link https://en.wikipedia.org/wiki/French_orthography | 'fra'} |