+++ src/guards/string/string-types.ts
@@ -44,6 +44,7 @@ export type StringRule =
 export type Alphabet =
     | 'basic-latin'
     | 'precomposed-latin'
+    | 'dan'
     | 'deu'
     | 'fin'
     | 'fra'