+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Add optional parameters 'fin', nld', and 'pol' to Tyr.string().isAlpha() method.
+-   Add optional parameters 'dan', 'fin', nld', and 'pol' to Tyr.string().isAlpha() method.
 
 ## [2.4.0] - 2021-06-03